refactor(#1526): extract semantic-consolidation coordinator (seam 5)
Extract the 366-line runSemanticConsolidation pipeline from
orchestrator.ts into SemanticConsolidationCoordinator
(orchestration/semantic-consolidation-coordinator.ts). This is the
5th orchestrator decomposition seam (issue #1526).

The coordinator owns the full semantic-consolidation lifecycle:
cluster discovery, LLM synthesis (operator-aware + legacy), canonical
write, source archival, index cleanup, and post-consolidation hooks
(Codex materialize, peer-profile reasoner).

Dependencies are injected via the Deps interface (config, getStorage,
fastLlm, embeddingFallback, removeContentHashForMemory callback,
saveContentHashIndexes callback) following the same pattern as the
existing compression-guideline and extraction-queue coordinators.

The orchestrator keeps a thin delegating method so all call sites
(runConsolidation, runSemanticConsolidationNow,
runSemanticConsolidationFanout) and existing tests remain stable.

Behavior-preserving: no logic changes. Dynamic imports in the original
method are converted to static imports in the coordinator module.

orchestrator.ts LOC: 20654 → 20311 (−343)
Ratchet baseline updated.

Chokepoints used: none new (delegation pattern only).
Not applicable: no new MCP tools, HTTP routes, or CLI commands.
